ID 타입 BigInteger에서 Long 변경 확인
S03P13A403-71 수정이 덜 된 내용 전부 수정 후 기능 테스트 완료

+++ backend/src/main/java/com/blsa/ezilog/dao/LikeCountDao.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.blsa.ezilog.model.like.LikeCount;
 
-public interface LikeCountDao extends JpaRepository<LikeCount, BigInteger> {
+public interface LikeCountDao extends JpaRepository<LikeCount, Long> {
 
     // 현재 이글에 내가 좋아요, 싫어요를 눌렀는지 체크
     @Query(value = "SELECT * FROM like_count WHERE uid=:uid AND type=:type AND post_no=:postno", nativeQuery = true)
